refactor application factory to utilize pytest.fixtures, update readme
Major refactor of the application factory in order to setup the app for ease of testing.  The app now has the ability to setup a test client with a test database via pytest.fixtures that match the features file scenario.  Updated readme with notes from where I got this information from.