Update readme.  Refactor app initialization.  Refactor test_db init
Updated readme to have an explanation on how to clone and get the app running.  Refactored the create_app function for initialization that creates a db file when the program is run instead of having to hop into a python repl.  Refactored init_database fixture as we don't need to initialize the db anymore.